feat: move config code to top level script
